Refactored PlansList.tsx - separated logic to a new file - usePlansListLogic.tsx, in PlansList.tsx left only the view.
