+++ src/components/PlansList/PlansList.tsx
@@ -1,13 +1,11 @@
-import React, {useEffect} from 'react';
-import {PlanEntity, Status} from 'types';
+import React from 'react';
+import {Status} from 'types';
 import {TbQuestionMark, TbX, TbHeartbeat, TbDotsVertical, TbUserCircle} from "react-icons/tb";
 import {IconContext} from "react-icons";
 import {Link} from "react-router-dom";
-import {apiUrl} from "../../config/api";
 import './PlansList.css';
 import {PlansListForm} from "./PlansListForm";
 import {MoonLoader} from "react-spinners";
-import {isDemoEnabled} from "../../helpers/env";
 import {DemoSign} from "../DemoSign/DemoSign";
 import {demoText} from "../../constants/demoText";
 import {ConfirmDeleteModal} from "../ConfirmDeleteModal/ConfirmDeleteModal";
@@ -23,142 +21,21 @@ export const PlansList = () => {
         plansList,
         isEdited,
         confirmDeletePlan,
-        planToDeleteId,
         isLoading,
-        setPlansList,
-        setIsEdited,
-        setConfirmDeletePlan,
-        setPlanToDeleteId,
-        setIsLoading,
+        addPlan,
+        editPlan,
+        handleUpdatePlan,
+        handleDeletePlan,
+        handleConfirmDelete,
+        handleCancelDelete,
     } = usePlansListLogic();
 
     const {
         informationModalIsOpen,
         demoModalIsOpen,
-        setInformationModalIsOpen,
-        setDemoModalIsOpen,
         closeModal,
         closeDemoModal,
     } = useModal();
-
-    useEffect(() => {
-        const abortController = new AbortController();
-
-        fetch(`${apiUrl}/api/add-plan/list`, {
-            method: 'GET',
-            signal: abortController.signal
-        }).then(res => res.json())
-            .then((plans) => {
-                setPlansList(plans);
-                setIsLoading(false);
-            })
-            .catch((error) => {
-                console.error("Wystąpił błąd podczas próby pobrania danych o planach treningowych:", error);
-                setIsLoading(false);
-            });
-
-        return () => {
-            try {
-                abortController.abort()
-            } catch {
-            }
-        };
-    }, [])
-
-    const addPlan = async (values: PlanEntity) => {
-        if (isDemoEnabled()) {
-            setDemoModalIsOpen(true); // Otwórz demoModal, jeśli demo jest włączone
-        } else if (values.name) {
-            const res = await fetch(`${apiUrl}/api/add-plan/list`, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify(values),
-            });
-
-            const data = await res.json();
-
-            setPlansList(list => [...list, data]);
-        } else {
-            setInformationModalIsOpen(true); // Otwórz informationModal, jeśli brakuje nazwy planu
-        }
-    };
-
-    const editPlan = async (values: PlanEntity, reset: () => void) => {
-
-        setIsEdited(false);
-
-        if (isDemoEnabled()) {
-            setDemoModalIsOpen(true);
-            reset();
-        } else if (values.name) {
-
-        const res = await fetch(`${apiUrl}/api/add-plan/list/${values.id}`, {
-            method: 'PUT',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify(values),
-        })
-
-        if (!res.ok) {
-            throw new Error('Wystąpił błąd podczas próby zaktualizowania nazwy planu.');
-        }
-
-        setIsEdited(true);
-
-        return await res.json();
-
-        } else {
-            setInformationModalIsOpen(true);
-            reset();
-        }
-    };
-
-    const handleUpdatePlan = (updatedPlan: PlanEntity) => {
-        setPlansList((plansList) =>
-            plansList.map((plan) =>
-                plan.id === updatedPlan.id ? updatedPlan : plan
-            )
-        );
-    };
-
-    const handleDeletePlan = async (planId: any) => {
-        if (isDemoEnabled()) {
-            setDemoModalIsOpen(true);
-            setPlanToDeleteId(planId);
-        } else {
-            setConfirmDeletePlan(true);
-            setPlanToDeleteId(planId);
-        }
-    };
-
-    const handleConfirmDelete = async () => {
-        if (isDemoEnabled()) {
-            closeDemoModal();
-        } else {
-            const res = await fetch(
-                `${apiUrl}/api/add-plan/list/${planToDeleteId}`,
-                {method: "DELETE"}
-            );
-            if ([400, 500].includes(res.status)) {
-                const error = await res.json();
-                alert(`Wystąpił błąd: ${error.message}`);
-                return;
-            }
-            setPlansList((plansList) =>
-                plansList.filter((plan) => plan.id !== planToDeleteId)
-            );
-            setConfirmDeletePlan(false);
-            setPlanToDeleteId(null);
-        }
-    };
-
-    const handleCancelDelete = () => {
-        setConfirmDeletePlan(false);
-        setPlanToDeleteId(null);
-    };
 
     if (isLoading || !plansList) {
         return (
